refactor components to its own files
